Move NewYear2021SaleBanner to a higher (shared) level
