+++ client/components/jetpack/new-year-2021-sale-banner/new-year-banner-2x.png
