+++ client/components/jetpack/new-year-2021-sale-banner/style.scss
