+++ client/jetpack-cloud/sections/pricing/header/index.tsx
@@ -13,13 +13,12 @@ import FormattedHeader from 'calypso/components/formatted-header';
 import OlarkChat from 'calypso/components/olark-chat';
 import config from 'calypso/config';
 import { preventWidows } from 'calypso/lib/formatting';
-import { useLocalizedMoment } from 'calypso/components/localized-moment';
 import { getJetpackCROActiveVersion } from 'calypso/my-sites/plans/jetpack-plans/abtest';
 import { Iterations } from 'calypso/my-sites/plans/jetpack-plans/iterations';
 
 // New Year 2021 promotion; runs from Jan 1 00:00 to Jan 18 23:59 UTC automatically.
 // Safe to remove on or after Jan 19.
-import NewYear2021SaleBanner from './new-year-2021-sale-banner';
+import NewYear2021SaleBanner from 'calypso/components/jetpack/new-year-2021-sale-banner';
 
 /**
  * Style dependencies
@@ -30,28 +29,6 @@ const Header: React.FC< Props > = ( { urlQueryArgs } ) => {
 	const identity = config( 'olark_chat_identity' );
 	const translate = useTranslate();
 	const iteration = useMemo( getJetpackCROActiveVersion, [] ) as Iterations;
-
-	const moment = useLocalizedMoment();
-	const showPromo = useMemo( () => {
-		const paramValue = urlQueryArgs?.[ 'newpack' ];
-
-		// If we're not forcing visibility via a query param,
-		// fall back to checking against the current browser time
-		if ( paramValue === undefined ) {
-			// Start at midnight UTC on Jan 1,
-			// and continue until 23:59:59 UTC on Jan 18
-			const promoStart = moment.utc( '2021-01-01' );
-			const promoEnd = moment.utc( '2021-01-19' );
-
-			return moment().isBetween( promoStart, promoEnd, 'second', '[)' );
-		}
-
-		// Show the promo if the `newpack` arg is present,
-		// as long as the arg value isn't '0' or 'false'.
-		return ! [ '0', 'false' ].some(
-			( v ) => v.localeCompare( paramValue, 'en', { sensitivity: 'base' } ) === 0
-		);
-	}, [ urlQueryArgs, moment ] );
 
 	const title =
 		{
@@ -76,7 +53,7 @@ const Header: React.FC< Props > = ( { urlQueryArgs } ) => {
 			{ identity && <OlarkChat { ...{ identity } } /> }
 			<JetpackComMasterbar />
 
-			{ showPromo && <NewYear2021SaleBanner /> }
+			<NewYear2021SaleBanner urlQueryArgs={ urlQueryArgs } />
 
 			<div className={ classNames( 'header', iteration ) }>
 				<FormattedHeader
